Upload whole project to GitHub.
After researching the best methods to document my code, I've finally decided that GitHub would be the best platform for me. Unfortunately, I've already created a large majority of the base code without proper documentation, and for that I apologize.

This project was started 2 months ago in the beginning of August 2022, but I needed a testable prototype by the last week of October 2022.  This development period was extremely crunched and I was the only developer on this project, which led me to neglect proper documentation of my work. However, I realize now that that shouldn't be an excuse for myself and that proper documentation is vital for any professional project.

Going forward, I hope to start better coding practices and documentation so that I can become an even better programmer.

Update README.md

Update README.md

Create README.md

Create main_window.py

The main window file contains the parent object to all window GUIs