Home Menu Creation
I removed the old optionmenu for switching between programs and replaced it with an new menu bar at the top of the window. I also added a new Home Window that the user will be directed to after they log in. Will add an option to log out of the program soon.