Added a new json file and reverse dict function
I thought of a new way to more easily store data format information. For now, I think the best method is to store the format information into an outside json file, that way it would be easier to update the code after the program is converted into an executable. To help aid this method, I added a new function called reverse_format_dict to the datatable object so that could easily reverse the format_dictionary when needed.