Added a textscroller to home window
I added a new widget to the program: the Text Scroller. The text scroller moves text from the right side of the screen to the left and loops back to the right side. Every time it loops, a new message will appear, of which there are about 100 different messages. These messages are stored in the tip.txt file in the Assets folder.

I used some old code for the Text Scroller that I created in a previous project iteration and optimized it for this project.