Added new json functionality for datatables
I created a new json file to hold all the data formatting information for the datatables. I figured that if changes were to be needed after product launch, it would be easier to change the datatables with the json file instead. I hear that other companies use xml files to store software gui information, but unfortunately, I don't have enough experience with those yet to justify implementation into the project.