Added new home image and clearance checks
I created a new function called clkearance_check  that helps determine if the user is allowed to view specific content of the code. This function was added to the the main window object so that very window has the capability of checking the user's access to the conten. i also aldded the company's logo to the home screen.