Major SQL fetching revamp
I decided to go forward with separating SQL queries from the being hardcoded in the program. I added the column names in data_format.json and created a reference to those columns in the code instead. Now, it should be easier to develop and debug the code going forward since I don't have to manually change the index of the data everytime I have to change the database structure.

I also added more quotes to tips.txt for the TextScroll widget on the Home window.